% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cat.R
\name{adjust_text_alignment}
\alias{adjust_text_alignment}
\title{Adjust text alignment}
\usage{
adjust_text_alignment(angle)
}
\arguments{
\item{angle}{Numeric, angle of text}
}
\description{
Adjust text alignment
}
\keyword{internal}
