% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-seq.R
\name{cat_dimplot}
\alias{cat_dimplot}
\title{Create a dimensionality reduction plot for categorical data}
\usage{
cat_dimplot(
  object,
  reduction = NULL,
  group_by = NULL,
  split_by = NULL,
  label = FALSE,
  raster = TRUE,
  show_legend = TRUE,
  show_axis = TRUE,
  show_border = TRUE,
  title = NULL,
  pt_size = 1,
  repel = TRUE,
  palette = "Paired",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing categorical data.}

\item{reduction}{The dimensionality reduction method. Default is NULL.}

\item{group_by}{The variable to group data by. Default is NULL.}

\item{split_by}{The variable to split data by. Default is NULL.}

\item{label}{Logical value indicating whether to show labels on the plot. Default is FALSE.}

\item{raster}{Logical value indicating whether to use rasterization for improved performance. Default is TRUE.}

\item{show_legend}{Logical value indicating whether to show the legend on the plot. Default is TRUE.}

\item{show_axis}{Logical value indicating whether to show the axis on the plot. Default is TRUE.}

\item{show_border}{Logical value indicating whether to show the panel and axis borders on the plot. Default is TRUE.}

\item{title}{The title for the plot. Default is NULL.}

\item{pt_size}{The size of the points on the plot. Default is 1.}

\item{repel}{Logical value indicating whether to use point repulsion to avoid overlapping labels. Default is TRUE.}

\item{palette}{The color palette to use for the plot. Default is "Paired".}

\item{...}{Additional parameters to be passed to the DimPlot() function in Seurat.}
}
\value{
A ggplot2 object containing the dimensionality reduction plot.
}
\description{
This function creates a dimensionality reduction plot for categorical data using Seurat and ggplot2. It allows for the selection of the reduction method, grouping, and splitting variables, as well as the visualization of labels, rasterization, and color palette. The cat_dimplot() function is intended to be used as a wrapper around Seurat's DimPlot() function.
}
\examples{
library(Seurat)
library(ggplot2)
data("pbmc_small")
pbmc_small <- CreateSeuratObject(counts = pbmc_small)
Idents(pbmc_small) <- pbmc_small$percent.mt > 10
cat_dimplot(object = pbmc_small, reduction = "tsne", group_by = "ident", palette = "Set1")

}
