% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-seq.R
\name{cat_dotplot}
\alias{cat_dotplot}
\title{Plot a dot plot with categorical groups}
\usage{
cat_dotplot(
  x,
  features,
  dot_scale = 4,
  col_min = -2.5,
  col_max = 2.5,
  assay = NULL
)
}
\arguments{
\item{x}{A Seurat object containing the data to plot.}

\item{features}{A character vector of feature names to plot.}

\item{dot_scale}{The size of the dots to plot.}

\item{col_min}{The minimum value for the color scale. Defaults to -2.5.}

\item{col_max}{The maximum value for the color scale. Defaults to 2.5.}
}
\value{
A ggplot2 object.
}
\description{
This function plots a dot plot with categorical groups using the DotPlot function from the Seurat package.
}
\examples{
cat_dotplot(x = mySeuratObject, features = c("CD3D", "CD8A", "CD4"))

}
