% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_modify_data.R
\name{cat_modify_data}
\alias{cat_modify_data}
\title{Modify factor levels of data}
\usage{
cat_modify_data(data, x, y, x_order = NULL, y_order = NULL)
}
\arguments{
\item{data}{A data frame to be modified.}

\item{x}{A column name of data frame to be adjusted factor levels.}

\item{y}{A column name of data frame to be adjusted factor levels.}

\item{x_order}{Optional. A character vector specifying how to adjust the
factor levels of the x column. It can be one of the following: "desc"
(arrange in descending order according to the mean value), "asc" (arrange
in ascending order according to the mean value), a vector of factor levels
in the desired order, or NULL (do not change).}

\item{y_order}{Optional. A character vector specifying how to adjust the
factor levels of the y column. It can be one of the following: "desc"
(arrange in descending order according to the mean value), "asc" (arrange
in ascending order according to the mean value), a vector of factor levels
in the desired order, or NULL (do not change).}
}
\value{
A modified data frame with adjusted factor levels.
}
\description{
This function is used to adjust the factor levels of data based on the
x_order and y_order inputs.
}
\examples{
library(catplot)
data(iris)
head(cat_modify_data(
  data = iris,
  x = Species,
  y = Sepal.Length,
  x_order =
    c("virginica", "versicolor", "setosa")
))
head(cat_modify_data(
  data = iris,
  x = Species,
  y = Sepal.Length,
  x_order = "desc"
))

}
