% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cat_save}
\alias{cat_save}
\title{Save a ggplot object to a file}
\usage{
cat_save(
  filename,
  plot = last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  raw_data = FALSE,
  ...
)
}
\arguments{
\item{filename}{A character string specifying the name of the file to save
the plot to}

\item{plot}{A ggplot object to be saved}

\item{device}{A character string specifying the graphics device to use}

\item{path}{A character string specifying the directory to save the file to}

\item{scale}{A numeric value specifying the scaling factor for the plot}

\item{width}{A numeric value specifying the width of the plot}

\item{height}{A numeric value specifying the height of the plot}

\item{units}{A character string specifying the units of the width and height
arguments}

\item{dpi}{A numeric value specifying the resolution of the plot}

\item{limitsize}{A logical value specifying whether to limit the size of the
plot to the device size}

\item{bg}{A character string specifying the background color of the plot}

\item{raw_data}{A logical value specifying whether to save the raw data used
to create the plot as a separate CSV file}

\item{...}{Additional arguments to be passed to ggsave}
}
\description{
This function saves a ggplot object to a file. It provides an option to save
the raw data used to create the plot as a separate CSV file.
}
