% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cat_set_path}
\alias{cat_set_path}
\title{Create a directory if it does not exist and return the path}
\usage{
cat_set_path(path)
}
\arguments{
\item{path}{A character string specifying the path to be created.}
}
\value{
A character string specifying the path.
}
\description{
This function takes a path as input and creates a directory if it does not
exist. It then returns the path.
}
\examples{
cat_set_path("~/my_folder")

}
