% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{cat_survival}
\alias{cat_survival}
\title{Plot survival probability}
\usage{
cat_survival(
  x,
  add_censor_mark = TRUE,
  add_pvalue = TRUE,
  add_confidence_interval = FALSE,
  palette = "status_1"
)
}
\description{
Plot survival probability
}
\examples{
library(ggsurvfit)
library(catplot)

x <- survfit2(Surv(time, status) ~ sex, data = df_lung)

x |> cat_survival()

x |> cat_survival(add_confidence_interval = TRUE)
}
\concept{others}
