% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-seq.R
\name{cat_vlnplot}
\alias{cat_vlnplot}
\title{Plot a violin plot with categorical groups}
\usage{
cat_vlnplot(
  object,
  features,
  pt_size = 0,
  sort = FALSE,
  group_by = NULL,
  split_by = NULL,
  show_legend = FALSE,
  angle_x = 0,
  aspect_ratio = 0.5
)
}
\arguments{
\item{object}{A Seurat object containing the data to plot.}

\item{features}{A character vector of feature names to plot.}

\item{pt_size}{The size of the points to plot.}

\item{sort}{Whether to sort the features by their mean expression or not.}

\item{group_by}{A character vector specifying the grouping variable. Defaults to "ident".}

\item{split_by}{A character vector specifying the splitting variable.}

\item{show_legend}{Whether to show the legend or not. Defaults to FALSE.}

\item{angle_x}{The angle of the x-axis labels. Defaults to 0.}

\item{aspect_ratio}{The aspect ratio of the plot. Defaults to 0.5.}
}
\value{
A ggplot2 object.
}
\description{
This function plots a violin plot with categorical groups using the VlnPlot function from the Seurat package.
}
\examples{
cat_vlnplot(object = mySeuratObject, features = c("CD3D", "CD8A", "CD4"))

}
