% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{catplot-package}
\alias{catplot}
\alias{catplot-package}
\title{catplot: Compute And Tity Plot Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A capable and tity data visualization tool.
}
\author{
\strong{Maintainer}: Songqi Duan \email{songqi.duan@outlook.com} (\href{https://orcid.org/0000-0002-0822-5883}{ORCID})

}
\keyword{internal}
