% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_ap}
\alias{create_ap}
\title{Construct aesthetic mappings or parameters for a plot}
\usage{
create_ap(color = NULL, fill = NULL, shape = NULL, alpha = NULL, size = NULL)
}
\arguments{
\item{color}{A character string specifying the color to be used in the plot}

\item{fill}{A character string specifying the fill color to be used in the plot}

\item{shape}{A character string specifying the shape to be used in the plot}

\item{alpha}{A character string specifying the alpha value to be used in the plot}

\item{size}{A character string specifying the size to be used in the plot}
}
\value{
A list containing the mapping and parameters for the plot
}
\description{
Create a mapping and parameters for a plot with a given \code{color} and
\code{fill}. The function checks if the color and fill arguments are valid
colors and, if so, adds them to the list of \code{params}. Otherwise, they
are added to the \code{mapping_props} list. Additional parameter can be
provided to customize the mapping and parameters.
}
