% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_new_project}
\alias{create_new_project}
\title{Create a new project directory structure}
\usage{
create_new_project(path)
}
\arguments{
\item{path}{A character string specifying the path to the project directory}
}
\value{
A character string specifying the path to the project directory
}
\description{
Create a new project directory structure
}
\examples{
create_new_project("~/my_project")
}
