% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{.get_brewer_palette}
\alias{.get_brewer_palette}
\title{Get a color palette from the RColorBrewer package}
\usage{
.get_brewer_palette(palette, n)
}
\arguments{
\item{palette}{A character value indicating the name of the RColorBrewer
palette to use}

\item{n}{An integer value specifying the number of colors to use from the
palette.}
}
\value{
A vector of colors from the selected RColorBrewer palette
}
\description{
This function returns a color palette from the RColorBrewer package by name,
with the option to specify the number of colors in the palette.
}
\examples{
palette <- catplot:::.get_brewer_palette("Blues", 6)

barplot(1:6, col = palette)
}
