% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{.get_ggsci_palette}
\alias{.get_ggsci_palette}
\title{Get a color palette from the ggsci package}
\usage{
.get_ggsci_palette(
  palette = c("npg", "aaas", "nejm", "lancet", "jama", "jco", "ucscgb", "d3",
    "locuszoom", "igv", "uchicago", "startrek", "tron", "futurama", "rickandmorty",
    "simpsons"),
  n
)
}
\arguments{
\item{palette}{A character value indicating the name of the ggsci palette to
use}

\item{n}{An integer value specifying the number of colors to use from the
palette. Defaults to the maximum number supported by the palette if not
specified.}
}
\value{
A vector of colors from the selected ggsci palette
}
\description{
This function returns a color palette from the ggsci package by name, with
the option to specify the number of colors in the palette.
}
\examples{
\dontrun{
palette <- catplot:::.get_ggsci_palette("npg", 6)

barplot(1:6, col = palette)
}
}
