% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.is_char_or_factor}
\alias{.is_char_or_factor}
\title{Check if an object is a character vector or a factor}
\usage{
.is_char_or_factor(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
A logical value indicating whether the object is a character vector
  or a factor
}
\description{
Check if a given object is a character vector or a factor.
This function is used internally by other functions.
}
\examples{
x <- c("a", "b", "c")
y <- factor(x)
catplot:::.is_char_or_factor(x)
catplot:::.is_char_or_factor(y)
}
