% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.is_color}
\alias{.is_color}
\title{Check if a string is a valid color value}
\usage{
.is_color(str)
}
\arguments{
\item{str}{A character string specifying the color value to be checked.}
}
\value{
A logical value indicating whether the input string is a valid color
  value.
}
\description{
This function takes a string and attempts to parse it as a color value using
the \code{col2rgb()} function. If the color value is valid, \code{TRUE} is
returned. Otherwise, \code{FALSE} is returned.
}
\examples{
catplot:::.is_color("red") # returns TRUE
catplot:::.is_color("#FFA500") # returns TRUE
catplot:::.is_color("fdafd") # returns FALSE
}
