% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{.is_color_palette}
\alias{.is_color_palette}
\title{Check if a palette name is valid}
\usage{
.is_color_palette(palette)
}
\arguments{
\item{palette}{A character vector containing the names of the color palettes
to check}
}
\value{
A logical value indicating whether the palette is valid
}
\description{
Check if a provided name or vector of names corresponds to a valid color
palette. Valid palettes include those from 'ggplot2' and 'RColorBrewer', as
well as some additional default palettes.
}
\examples{
catplot:::.is_color_palette("Paired") # TRUE
catplot:::.is_color_palette(c("foo")) # FALSE
}
