% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate a color palette}
\usage{
generate_colors(
  palette = NULL,
  n = NULL,
  color_names = NULL,
  modification = NULL,
  view_palette = TRUE,
  view_labels = TRUE
)
}
\arguments{
\item{palette}{A character string specifying the name of the color palette to
use. The default value is "default". Valid options include "default",
"hue", "grey", "gray", "brewer", and "ggsci".}

\item{n}{An integer specifying the number of colors to generate. The default
value is NULL, which generates a default number of colors based on the
selected palette.}

\item{color_names}{A character vector specifying names for the individual
colors in the palette. The length of this vector should be the same as the
number of colors in the palette. The default value is NULL, which does not
assign names to the colors.}

\item{modification}{A character string specifying the modification to apply
to the colors. The default value is "go_lighter". Valid options include
"go_darker", "go_lighter", and "go_both_ways".}

\item{view_palette}{A logical value indicating whether to display the
generated color palette. The default value is TRUE.}

\item{view_labels}{A logical value indicating whether to display labels for
the colors in the generated palette. The default value is TRUE.}
}
\value{
A vector of colors, with each element representing a color in the
  generated palette.
}
\description{
This function generates a color palette based on the input parameters.
}
\examples{
generate_colors(palette = "default", n = 5)
generate_colors(palette = "Paired", n = 8, view_palette = FALSE)
generate_colors(
  palette = "npg",
  n = 3,
  color_names = c("red", "blue", "green"),
  view_labels = FALSE
)
}
