% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_boxplot.R
\name{geom_boxplot_internal}
\alias{geom_boxplot_internal}
\title{Create a box plot with customizable color or fill}
\usage{
geom_boxplot_internal(color = NULL, fill = NULL, ...)
}
\arguments{
\item{color}{A character string representing the color to use for the box
plot outline}

\item{fill}{A character string representing the color to use for the box plot
fill}

\item{...}{Additional arguments to be passed to
\code{\link[ggplot2]{geom_boxplot}}}
}
\value{
A ggplot object with a box plot
}
\description{
This function creates a box plot using the \code{geom_boxplot()} function
from the \code{ggplot2} package. The color or fill aesthetic can be
customized using the \code{color} or \code{fill} parameter, respectively.
}
\examples{
library(catplot)
library(ggplot2)
# Create a box plot with red outline
geom_boxplot_internal(color = "red")

# Create a box plot with green fill
geom_boxplot_internal(fill = "green")

# Create a box plot with custom data
data <- data.frame(x = rep(c("A", "B"), each = 10), y = rnorm(20))
ggplot(data, aes(x, y)) +
  geom_boxplot_internal(color = "blue", fill = "yellow")
}
