% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_char_or_factor}
\alias{is_char_or_factor}
\title{Check if a variable is a character or factor}
\usage{
is_char_or_factor(x)
}
\arguments{
\item{x}{A variable to be checked.}
}
\value{
A logical value indicating whether the variable is a character or factor.
}
\description{
This function checks if a variable is a character or factor.
}
\examples{
library(catplot)
is_char_or_factor("hello")
is_char_or_factor(factor("world"))

}
