% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_color}
\alias{is_color}
\title{Check if a string is a valid color}
\usage{
is_color(str)
}
\arguments{
\item{str}{A character string representing a color}
}
\value{
A logical value indicating whether the input is a valid color
}
\description{
This function attempts to convert a string to an RGB value using the
\code{col2rgb()} function from the \code{grDevices} package. If the RGB value
is not empty, the string is considered a valid color.
}
\examples{
library(catplot)
is_color("red") # TRUE
is_color("#00FF00") # TRUE
is_color("not a color") # FALSE
}
