% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_catplot}
\alias{scale_color_catplot}
\title{Scale color for categorical plots}
\usage{
scale_color_catplot(
  palette = "Paired",
  n = NULL,
  color_names = NULL,
  modification = "go_lighter",
  direction = 1
)
}
\arguments{
\item{palette}{A character string specifying the name of the color palette to
use. The default value is "Paired". Valid options include any palette name
recognized by the \code{\link{generate_colors}} function.}

\item{n}{An integer specifying the number of colors to generate. The default
value is NULL, which generates a default number of colors based on the
selected palette.}

\item{color_names}{A character vector specifying names for the individual
colors in the palette. The length of this vector should be the same as the
number of colors in the palette. The default value is NULL, which does not
assign names to the colors.}

\item{modification}{A character string specifying the modification to apply
to the colors. The default value is "go_lighter". Valid options include
"go_darker", "go_lighter", and "go_both_ways".}

\item{direction}{An integer specifying the direction in which to order the
colors. The default value is 1, which orders the colors from light to dark.
A value of -1 orders the colors from dark to light.}
}
\value{
A manual scale for color aesthetics using the generated colors.
}
\description{
This function generates colors using the \code{\link{generate_colors}}
function and applies them to a manual scale for color aesthetics.
}
\examples{
scale_color_catplot(palette = "Set1", n = 3)
scale_color_catplot(
  palette = "Dark2",
  n = 3,
  color_names = c("A", "B", "C"),
  direction = -1
)
}
