% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_modify_plot.R
\name{theme_cat}
\alias{theme_cat}
\title{Custom theme for catplot}
\usage{
theme_cat(
  font_size = 8,
  linewidth = 0.5,
  aspect_ratio = NULL,
  frame = "closed",
  show_panel_grid = "both_not",
  show_title = "both",
  show_text = "both",
  text_italic = "both_not",
  show_ticks = "both",
  ticks_length = 4,
  legend_position = "right",
  legend_direction = NULL
)
}
\arguments{
\item{font_size}{Numeric, font size (default: 8)}

\item{linewidth}{Numeric, line width (default: 0.5)}

\item{aspect_ratio}{Numeric, aspect ratio of the plot (default: NULL)}

\item{frame}{Character, frame type, "closed" or "open" (default: "closed")}

\item{show_panel_grid}{Character, panel grid visibility (default: "both_not")}

\item{show_title}{Character, axis title visibility (default: "both")}

\item{show_text}{Character, axis text visibility (default: "both")}

\item{text_italic}{Character, axis text italicization (default: "both_not")}

\item{show_ticks}{Character, axis tick visibility (default: "both")}

\item{ticks_length}{Numeric, tick length (default: 4)}

\item{legend_position}{Character, position of legend (default: "right")}

\item{legend_direction}{Character, direction of legend (default: NULL)}
}
\value{
A ggplot2 theme
}
\description{
Custom theme for catplot
}
\examples{
library(catplot)
library(ggplot2)

p <- ggplot(mpg, aes(x = manufacturer, y = hwy)) +
  geom_point()
p + theme_cat()
}
