% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{cat_boxplot}
\alias{cat_boxplot}
\title{Create a boxplot with customized options using ggplot2}
\usage{
cat_boxplot(
  data,
  x,
  y,
  fill = NULL,
  width = 0.6,
  fatten = 1,
  show_legend = FALSE,
  aspect_ratio = NULL,
  flip = FALSE,
  frame = "closed",
  font_size = 8,
  show_panel_grid = NULL,
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL,
  show_title = "y",
  show_ticks = "both",
  show_text = "both",
  text_italic = NULL,
  x_text_angle = NULL,
  y_text_angle = NULL,
  legend_title = NULL,
  legend_position = "right",
  legend_direction = NULL,
  comparisons = NULL,
  test = "wilcox.test",
  step_increase = 0.1,
  map_signif_level = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{fill}{The variable to be used for filling the boxes. Defaults to NULL.}

\item{width}{The width of the boxes. Defaults to 0.6.}

\item{fatten}{The amount by which to fatten the boxes. Defaults to 1.}

\item{show_legend}{Logical indicating whether to show the legend. Defaults to FALSE.}

\item{aspect_ratio}{numeric; ratio of x to y axis length}

\item{flip}{Logical indicating whether to flip the axes of the plot. Defaults to FALSE.}

\item{frame}{character; indicates whether the plot frame should be "open" or
"closed"}

\item{font_size}{numeric; base size of text elements in the theme}

\item{show_panel_grid}{character; indicates which panel grid to show, "x",
"y", "both", or "both_not"}

\item{plot_title}{Title of the plot.}

\item{x_axis_title}{Title of the x-axis.}

\item{y_axis_title}{Title of the y-axis.}

\item{show_title}{character; indicates which axis title to show, "x", "y",
"both", or "both_not"}

\item{show_ticks}{character; indicates which axis ticks to show, "x", "y",
"both", or "both_not"}

\item{show_text}{character; indicates which axis text to show, "x", "y",
"both", or "both_not"}

\item{text_italic}{character; indicates which axis text to italicize, "x",
"y", "both", or "both_not"}

\item{x_text_angle}{Angle of the x-axis labels.}

\item{y_text_angle}{Angle of the y-axis labels.}

\item{legend_title}{Title of the legend.}

\item{legend_position}{Position of the legend, can be "left", "right", "top", "bottom", or NULL.}

\item{legend_direction}{Direction of the legend, can be "horizontal" or "vertical".}

\item{comparisons}{List of comparison groups for adding significance labels.}

\item{test}{Type of statistical test to use for calculating significance. Defaults to "wilcox.test".}

\item{step_increase}{Step increase for adjusting the significance label position.}

\item{map_signif_level}{Logical indicating whether to map the significance level to asterisks.}
}
\value{
A ggplot2 object representing the boxplot with customized options.
}
\description{
This function creates a boxplot using ggplot2 with options for modifying the appearance of the plot.
}
\examples{
library(catplot)
library(ggplot2)
library(dplyr)

data("iris")
iris \%>\%
  mutate(group = if_else(log2(Petal.Length + 1) > 1.5,
    "high", "low"
  )) \%>\%
  cat_boxplot(
    x = Species,
    y = log2(Petal.Width + 1),
    aspect_ratio = 1,
    show_panel_grid = "both",
    fill = group,
    legend_title = "ee",
    legend_position = "top",
    x_text_angle = 45
  )
p <- iris |>
  cat_boxplot(
    x = Species, y = Sepal.Length,
    aspect_ratio = 2, frame = "open",
    x_text_angle = 45,
    comparisons = list(c("versicolor", "virginica"))
  ) + scale_fill_brewer(palette = "Paired")
p
}
