% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{cat_modify_plot}
\alias{cat_modify_plot}
\title{Modify a ggplot2 plot with customized options and add significance labels}
\usage{
cat_modify_plot(
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL,
  x_text_angle = NULL,
  y_text_angle = NULL,
  legend_title = NULL,
  comparisons = NULL,
  test = "wilcox.test",
  step_increase = 0.1,
  map_signif_level = TRUE,
  aspect_ratio = NULL,
  frame = "closed",
  font_size = 8,
  show_panel_grid = NULL,
  show_title = "both",
  show_ticks = "both",
  show_text = "both",
  text_italic = NULL,
  legend_position = "right",
  legend_direction = NULL
)
}
\arguments{
\item{plot_title}{Title of the plot.}

\item{x_axis_title}{Title of the x-axis.}

\item{y_axis_title}{Title of the y-axis.}

\item{x_text_angle}{Angle of the x-axis labels.}

\item{y_text_angle}{Angle of the y-axis labels.}

\item{legend_title}{Title of the legend.}

\item{comparisons}{List of comparison groups for adding significance labels.}

\item{test}{Type of statistical test to use for calculating significance. Defaults to "wilcox.test".}

\item{step_increase}{Step increase for adjusting the significance label position.}

\item{map_signif_level}{Logical indicating whether to map the significance level to asterisks.}

\item{aspect_ratio}{numeric; ratio of x to y axis length}

\item{frame}{character; indicates whether the plot frame should be "open" or
"closed"}

\item{font_size}{numeric; base size of text elements in the theme}

\item{show_panel_grid}{character; indicates which panel grid to show, "x",
"y", "both", or "both_not"}

\item{show_title}{character; indicates which axis title to show, "x", "y",
"both", or "both_not"}

\item{show_ticks}{character; indicates which axis ticks to show, "x", "y",
"both", or "both_not"}

\item{show_text}{character; indicates which axis text to show, "x", "y",
"both", or "both_not"}

\item{text_italic}{character; indicates which axis text to italicize, "x",
"y", "both", or "both_not"}

\item{legend_position}{Position of the legend, can be "left", "right", "top", "bottom", or NULL.}

\item{legend_direction}{Direction of the legend, can be "horizontal" or "vertical".}
}
\value{
A list of ggplot2 options for modifying a plot and adding significance labels.
}
\description{
Modify a ggplot2 plot with customized options and add significance labels
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = factor(cyl), y = mpg)) +
  geom_boxplot() +
  cat_modify_plot(plot_title = "My Plot Title",
                x_axis_title = "X Axis Title",
                y_axis_title = "Y Axis Title",
                test = "t.test",
                comparisons = list(c("4", "6"), c("4", "8")),
                aspect_ratio = 1)
p

}
