% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cat.R
\name{theme_cat}
\alias{theme_cat}
\title{Custom theme for catplot}
\usage{
theme_cat(
  base_font_size = 8,
  font_family = NULL,
  linewidth = 0.5,
  aspect_ratio = NULL,
  frame = "open",
  show_panel_grid = "none",
  show_title = "both",
  show_text = "both",
  text_italic = "none",
  show_ticks = "both",
  ticks_length = 4,
  show_legend_title = TRUE,
  legend_position = "right",
  legend_position_inside = NULL,
  legend_direction = NULL,
  x_text_angle = 0,
  ...
)
}
\arguments{
\item{base_font_size}{Numeric, font size (default: 8)}

\item{font_family}{Character, font family (default: NULL)}

\item{linewidth}{Numeric, line width (default: 0.5)}

\item{aspect_ratio}{Numeric, aspect ratio of the plot (default: NULL)}

\item{frame}{Character, frame type, "none","closed" or "open" (default:
"closed")}

\item{show_panel_grid}{Character, panel grid visibility (default: "none")}

\item{show_title}{Character, axis title visibility (default: "both")}

\item{show_text}{Character, axis text visibility (default: "both")}

\item{text_italic}{Character, axis text italicization (default: "none")}

\item{show_ticks}{Character, axis tick visibility (default: "both")}

\item{ticks_length}{Numeric, tick length (default: 4)}

\item{show_legend_title}{Logical, show legend title (default: TRUE)}

\item{legend_position}{Character, position of legend (default: "right")}

\item{legend_position_inside}{Logical, position legend inside the plot
(default: NULL)}

\item{legend_direction}{Character, direction of legend (default: NULL)}

\item{x_text_angle}{Numeric, angle of x-axis text (default: 0)}

\item{...}{Additional arguments}
}
\value{
A ggplot2 theme

A ggplot2 theme
}
\description{
Custom theme for catplot
}
\examples{
library(catplot)
library(ggplot2)

p <- iris |>
  ggplot(aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(color = Species))
p
# Axis title
p + theme_cat(show_title = "y")

# Axis text
p + theme_cat(
  text_italic = "x",
  show_text = "x"
)

# Axis ticks
p + theme_cat(show_ticks = "y")

# Panel grid
p + theme_cat(show_panel_grid = "both")

# Frame
p + theme_cat(
  frame = "closed", aspect_ratio = 1,
  show_panel_grid = "both",
  show_title = "y"
)
}
\concept{theme}
