% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cat}
\alias{theme_cat}
\title{Custom theme for catplot function}
\usage{
theme_cat(
  aspect_ratio = NULL,
  frame = "closed",
  show_panel_grid = NULL,
  show_title = "both",
  show_ticks = "both",
  show_text = "both",
  text_italic = NULL,
  legend_position = "right",
  legend_direction = NULL,
  base_size = 8,
  base_family = "",
  base_line_size = base_size/16,
  base_rect_size = base_size/16,
  ...
)
}
\arguments{
\item{aspect_ratio}{numeric; ratio of x to y axis length}

\item{frame}{character; indicates whether the plot frame should be "open" or
"closed"}

\item{show_panel_grid}{character; indicates which panel grid to show, "x",
"y", "both", or "both_not"}

\item{show_title}{character; indicates which axis title to show, "x", "y",
"both", or "both_not"}

\item{show_ticks}{character; indicates which axis ticks to show, "x", "y",
"both", or "both_not"}

\item{show_text}{character; indicates which axis text to show, "x", "y",
"both", or "both_not"}

\item{text_italic}{character; indicates which axis text to italicize, "x",
"y", "both", or "both_not"}

\item{legend_position}{character; indicates the position of the legend,
"right", "left", "bottom", or "top"}

\item{legend_direction}{character; indicates the direction of the legend, "h"
or "horizontal", "v" or "vertical", or "default"}

\item{base_size}{numeric; base size of text elements in the theme}

\item{base_family}{character; base font family of text elements in the theme}

\item{base_line_size}{numeric; base line size of elements in the theme}

\item{base_rect_size}{numeric; base line size of rect in the theme}

\item{...}{additional arguments passed to \code{\link[ggplot2:theme]{theme}}}
}
\value{
a ggplot2 theme object
}
\description{
This is a custom theme for the catplot function in the catplot package. It
sets the visual appearance of the plot elements, such as axis, panel, legend,
and title, to match the default style of catplot plots.
}
\examples{
library(catplot)
library(ggplot2)

data("iris")
ggplot(data = iris, aes(x = Species, y = Sepal.Length)) +
geom_boxplot() +
theme_cat(
  aspect_ratio = 1,
  show_panel_grid = "both",
  show_title = "y"
)
}
